// open-banking-react-apps/lib/auth/src/types.ts
/**
 * Copyright (c) 2023, WSO2 LLC. (https://www.wso2.com).
 *
 * WSO2 LLC. licenses this file to you under the Apache License,
 * Version 2.0 (the "License"); you may not use this file except
 * in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing,
 * software distributed under the License is distributed on an
 * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
 * KIND, either express or implied. See the License for the
 * specific language governing permissions and limitations
 * under the License.
 */

export type AuthConfig = {
    accessTokenCookieName: string;
    refreshTokenCookieName: string;
    idTokenCookieName: string;
    baseUrl: string;
}

export type AuthProviderProps = {
    config: AuthConfig;
    children: React.ReactNode;
}

export type AuthState = {
    isAuthenticated?: boolean;
    isLoading?: boolean;
    username?: string;
    clientId?: string;
    role?: string;
}

export type AuthContextType = {
    state?: AuthState;
    setState?: (state: AuthState) => void;
    signIn?: () => void;
    signOut?: () => void;
    isAuthenticated?: boolean;
    getIDToken?: () => string;
    getDecodedIDToken?: () => { [key: string]: string };
    getPartialAccessToken?: () => string;
    getPartialRefreshToken?: () => string;
}

export type SecureRoutesProps = {
    redirectPath: string
}